Java_Kursova-main/financeapp/src/main/java/com/example/financeapp/service/BudgetService.java
package com.example.financeapp.service;
import com.example.financeapp.model.Budget;
import com.example.financeapp.repository.BudgetRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@RequiredArgsConstructor
public class BudgetService {

    private final BudgetRepository budgetRepository;

    public Budget addBudget(Budget budget) {
        return budgetRepository.save(budget);
    }

    public List<Budget> getUserBudgets(Long userId) {
        return budgetRepository.findByUserId(userId);
    }

    public void deleteBudget(Long id) {
        budgetRepository.deleteById(id);
    }
}